+++ frontend/src/components/Layout.test.tsx
@@ -34,7 +34,7 @@ describe("Layout", () => {
       </QueryClientProvider>,
     );
     expect(screen.getByText("Torneios")).toBeInTheDocument();
-    expect(screen.getByText("Ranking FP")).toBeInTheDocument();
+    expect(screen.getByText("Ranking Fourse Points")).toBeInTheDocument();
     expect(screen.queryByText("Premiação")).not.toBeInTheDocument();
     expect(screen.queryByText("Alterar senha")).not.toBeInTheDocument();
     expect(screen.getByText("Powered by")).toBeInTheDocument();